% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msvs.r
\name{get_library_path}
\alias{get_library_path}
\title{Gets the content of the LIBRARY_PATH environment variable}
\usage{
get_library_path(dirsep = "/", do_cat = TRUE)
}
\arguments{
\item{dirsep}{directory separator to expect/use for the retrieved value}

\item{do_cat}{use the `cat` function to output to stdout rather than return the character vector}
}
\value{
a character vector if do_cat=FALSE, nothing otherwise (default)
}
\description{
Gets the content of the LIBRARY_PATH environment variable
}
