% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msvs.r
\name{get_location}
\alias{get_location}
\title{Gets the location on disk of this package}
\usage{
get_location()
}
\description{
Gets the location on disk of this package. Usage is to locate a root directory for other specialized functions in this package.
}
