% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msvs.r
\name{get_makevars_win_template}
\alias{get_makevars_win_template}
\title{Gets the location of a template Makevars.win.in}
\usage{
get_makevars_win_template(dirsep = "/", do_cat = TRUE)
}
\arguments{
\item{dirsep}{directory separator to expect/use for the retrieved value}

\item{do_cat}{use the `cat` function to output to stdout rather than return the character vector}
}
\value{
a character vector if do_cat=FALSE, nothing otherwise (default)
}
\description{
Gets the location of a template Makevars.win.in
}
