% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msvs.r
\name{get_msbuild_exe_path}
\alias{get_msbuild_exe_path}
\title{Gets the location of msbuild.exe}
\usage{
get_msbuild_exe_path(dirsep = "/", do_cat = TRUE)
}
\arguments{
\item{dirsep}{directory separator to expect/use for the retrieved value}

\item{do_cat}{use the `cat` function to output to stdout rather than return the character vector}
}
\value{
a character vector if do_cat=FALSE, nothing otherwise (default)
}
\description{
Gets the location of msbuild.exe
}
