% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msvs.r
\name{get_win_cp_cmd}
\alias{get_win_cp_cmd}
\title{Gets the location of msvs's command line file copy utility}
\usage{
get_win_cp_cmd(dirsep = "/", do_cat = TRUE)
}
\description{
Gets the location of msvs's command line file copy utility
}
