% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msvs.r
\name{has_visualstudio}
\alias{has_visualstudio}
\title{Gets the location of msbuild.exe}
\usage{
has_visualstudio(major_version = 14, minor_version = 0)
}
\arguments{
\item{dirsep}{directory separator to expect/use for the retrieved value}
}
\value{
a character vector if do_cat=FALSE, nothing otherwise (default)
}
\description{
Gets the location of msbuild.exe
}
